Move readImageFromPath into UIStarter as static
I've been converting these methods to be non-static, but readImageFromPath is called by FileMoveIcon, and there's really no good way to get it a pointer to the UIStareter.

So, keep it static in UIStarter, too.  With this commit, we can eliminate UIUtils.
